+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.11.0 (not released yet)
 
-* The initial release.
+* The initial release, compatible with Sponge 1.11.x..